refactor: create tokens folder
