+++ src/pages/login/ui/LoginPage.tsx
@@ -5,7 +5,7 @@ import { Box, Grid } from "@mui/material";
 import { gridContainerProps, gridItemProps, gridHeaderProps } from "./style";
 import { useNavigate } from "react-router-dom";
 import { Paths } from "../../../shared/constants/paths";
-import { customerTokenCache } from "../../../shared/api/tokens";
+import { customerTokenCache } from "../../../shared/api/";
 
 export function LoginPage() {
   const cachedToken = customerTokenCache.get().token;
