+++ src/shared/api/index.ts
@@ -1,1 +1,2 @@
 export { createCustomer, loginCustomer } from "./customers";
+export { customerTokenCache } from "./tokens/tokens";
